Implement a new Hand class with better structure
