+++ src/scripts/card.js
@@ -1,0 +1,8 @@
+export default class Card {
+  constructor(name, suit, value, code) {
+    this.name = name
+    this.suit = suit
+    this.value = value
+    this.code = code
+  }
+}
