--- src/scripts/hand-ranker.js
@@ -1,12 +1,0 @@
-export const containsPair = (cards) => {
-  if (cards.length < 2) { return false }
-  let valuesCount = {}
-  for (const card of cards) {
-    if (valuesCount[card.value] === 1) {
-      return true
-    } else {
-      valuesCount[card.value] = 1
-    }
-  }
-  return false
-}
